mir: Create MirVisitor/MirVisitorMut traits
Create two traits to better handle walking the MIR tree.

MirVisitor will not mutate the tree, MirVisitorMut can mutate the
tree, either directly or via the return value of a visitor function
with the MVMAction enum.

Apply MirVisitor/MirVisitorMut to used_labels, defined_labels and
collapse_cmp, though not cull_fallthrough_jumps as it is not really
and improvement there.