dataflow: Inline uses of fp, where possible
armv8 uses fp and sp fairly interchangeably, which
can be confusing to the reader, as the same location
will be accessed differently. Therefore, try to
replace all references to fp with the equivalent sp
case, where it is possible.

To help with the readability of the output, also
add reduce_binops, which reduces e.g. (sp + 1) + 2
to sp + 3.